pyproject.toml
[tool.poetry]
name = "elasticsearch-rag-tutorial"
version = "0.1.0"
description = ""
authors = ["lizozom <lizka.k@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
kagglehub = "^0.3.7"
kaggle = "^1.6.17"
pandas = "^2.2.3"
cohere = "^5.13.11"
elasticsearch = "^8.17.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
